DISPATCH-1645: Use uint64_t for max message size at run time
Policy settings and statistics were using unsigned 64 bit values.
Runtime code was demoting them to signed 32 bit values.

The 1.12 implementation works correctly only for max message
size limits less than 2^31. This includes value up to 2,147,483,647.

+++ include/qpid/dispatch/server.h
@@ -604,7 +604,7 @@ bool qd_connection_strip_annotations_in(const qd_connection_t *c);
 
 void qd_connection_wake(qd_connection_t *ctx);
 
-int qd_connection_max_message_size(const qd_connection_t *c);
+uint64_t qd_connection_max_message_size(const qd_connection_t *c);
 
 /**
  * @}
